==> 周边人口/2.1 Crawl_ID_Lianjia.ipynb <==
from bs4 import BeautifulSoup as bs
import urllib
import requests
import pandas as pd
import re
import time

data = pd.read_csv("./Amap_info.csv")
# 删除有地址坐标但没有小区名称的数据
data = data[data['高德名称'].notnull()]
#删除名称中含括号内容的数据，否则在链家网搜获效率很低
data['高德名称'] = data['高德名称'].str.replace(r'\(.*?\)','')
data.drop(data.columns[0], axis=1, inplace=True)
data = data.reset_index(drop = True)
data.head()

# Select a random agent IP adress, and dim the headers
proxy = {'https': '102.164.212.54:8080'}
headers = {'User-Agent':
          "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36 Edg/84.0.522.61"}

new_data = pd.DataFrame(columns = ['高德ID','链家ID','高德名称','链家名称'])
# 在链家网找小区页面，根据高德提供的小区名称，爬取小区对应在链家网的名称以及内部编号
for i in range(len(data)):
    # 连接搜索页获得源码
    quote_name = urllib.parse.quote(data['名称'][i])                                #不支持中英文url地址，因此需要转换
    url_search = 'https://sh.lianjia.com/xiaoqu/fengxian/rs{}'.format(quote_name)
    response_search = requests.get(url_search, headers = headers)
    soup_search = bs(response_search.text, 'html.parser')
    # 查找搜索结果的数量
    search_result_txt = soup_search.find(class_ = "total fl").get_text()
    search_result_num = int(re.findall(r'\d+',search_result_txt)[0])
    # 根据不同的搜索结果分别处理：    
    ## 假如根据高德的小区名称搜索的到结果，可能唯一也可能不唯一
    if search_result_num != 0:
        info_content = soup_search.find(class_ = 'listContent').find_all('li')
        for j in range(search_result_num):
            id_lianjia = info_content[j]['data-housecode']
            name_lianjia = info_content[j].find(class_ = 'title').find('a').get_text()                        
            info_dict = {'高德ID' : data['ID'][i],
                         '链家ID' : id_lianjia,
                         '高德名称' : data['名称'][i],
                         '链家名称' : name_lianjia }
            new_data = new_data.append(info_dict, ignore_index=True)
    ## 假如搜索不到结果
    else:
        continue
    time.sleep(1)

new_data.head()

new_data = new_data.drop_duplicates('链家ID').reset_index(drop=True)
new_data.head()

new_data.to_csv("./Lianjia_info_partial.csv", encoding='utf_8_sig')
