==> 周边人口/2.2 Craw_DetailInfo_Lianjia.ipynb <==
from bs4 import BeautifulSoup as bs
from collections import defaultdict
import urllib
import requests
import pandas as pd
import re
import time

new_data = pd.read_csv("./Lianjia_info_partial.csv")
new_data.drop(new_data.columns[0], axis=1, inplace=True)
new_data = new_data.reset_index(drop = True)
new_data.head()

# Select a random agent IP adress, and dim the headers
proxy = {'https': '102.164.212.54:8080'}
headers = {'User-Agent':
          "Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/84.0.4147.125 Safari/537.36 Edg/84.0.522.61"}

# 按照链家网的小区ID编号搜索小区详细信息
resid_table = pd.DataFrame()
for resid_id in new_data['链家ID']:         
    # 构造小区网页地址
    url_resid = 'https://sh.lianjia.com/xiaoqu/{}'.format(resid_id)
    # 访问
    response = requests.get(url_resid, headers = headers)
    # soup一下美美的
    soup = bs(response.text, 'html.parser')
    # 信息截取
    # 网页源码中 label_list 和 content_list 两个关键内容
    label_list = soup.find_all(class_ = 'xiaoquInfoLabel')
    content_list = soup.find_all(class_ = 'xiaoquInfoContent')
    # 将抓到的信息存进DataFrame
    info_dict = defaultdict(list) #构造一个空字典且将其实例化的方法，便于append
    info_dict['链家ID'].append(resid_id)
    for i,j in zip(label_list,content_list):
        info_dict[i.get_text()].append(j.get_text())
    resid_table = resid_table.append(pd.DataFrame(info_dict),ignore_index=True)
    time.sleep(1)

resid_table.head()

resid_table.to_csv('./Lianjia_info_all.csv',encoding='utf_8_sig')
