==> 周边人口/3. Data_Cleaning.ipynb <==
import pandas as pd
data1 = pd.read_csv('./Amap_info.csv')
data2 = pd.read_csv('./Lianjia_info_partial.csv')
data3 = pd.read_csv("./Lianjia_info_all.csv")

data1.head()

data1.drop(data1.columns[0], axis=1, inplace=True)
#data1 = data1.set_index('高德ID')
data1.head()

data2.head()
#data_amap.drop(data_amap.columns[0], axis=1, inplace=True)
#data_amap = data_amap.reset_index(drop = True)
#data_amap

data2.drop(data2.columns[0], axis=1, inplace=True)
#data2 = data2.set_index(['高德ID','链家ID'])
data2.head()
#data2里面包含重复的索引，不能用pd.concat方法合并两张表

data1_2 = pd.merge(data1,data2,on='高德ID',how='inner')
data1_2.head()

data3.head()

data3.drop(data3.columns[0], axis=1, inplace=True)
data3 = data3.set_index('链家ID')
data3.head()

# 表1表2合并
data1_2 = data1_2.reset_index(drop=True).set_index('链家ID')
data1_2.head()

# 再合并表3的信息
data =data1_2.join(data3)
data=data.reset_index()
data.head()

data = data.drop(columns=['高德名称_y'],axis=1)

data.rename(columns={'高德名称_x':'高德名称'},inplace=True)
data.head()

# 去除空值的行
data = data[data['高德名称'].notnull() & data['楼栋总数'].notnull() & data['房屋总数'].notnull()]
data.head()

# 去除重复的行
data = data.drop_duplicates(subset=['链家ID']).reset_index()
data.head()

# 提取数值
data['楼栋总数'] = data['楼栋总数'].str.extract(r'(\d+)').astype('int')
data['房屋总数'] = data['房屋总数'].str.extract(r'(\d+)').astype('int')
data.head()

data.to_csv("./Data.csv", encoding='utf_8_sig')
