==> 周边人口/4.3 BubbleMap_baidu.ipynb <==
import requests
from requests.exceptions import ReadTimeout, ConnectTimeout
import pandas as pd

data = pd.read_csv('./data00.csv')
data.head()

with open('./Baidu_AK.txt','r') as f:
    AK_str = f.read()
    
# 高德火星坐标转换为百度坐标系
def coordconv(coord_str):
    # 要求原始的coord_str是经度在前纬度在后
    base =  "http://api.map.baidu.com/geoconv/v1/"
    parameters = {"coords": coord_str,
                  "ak":  AK_str,
                  "from": 3,
                  "to": 5
                 }
    try:
        response = requests.get(base, parameters, timeout=2)
        if response.status_code == 200:
            answer = response.json()
            loc_dict = answer['result'][0]
            lon = loc_dict.get('x')
            lat = loc_dict.get('y')
            loc = [lon,lat]
        else:
            pass
    except (ReadTimeout, ConnectTimeout):
        pass
    return loc

test = coordconv('121.480515,30.912384')
test

"""
I  .百度web API只能通过在线工具生成静态图，无法通过python调用生成

II .百度的pyecharts地理散点图可参考：
1. https://blog.csdn.net/zengbowengood/article/details/104695205
2. https://www.cnblogs.com/zhuwjwh/p/12095476.html
3. https://gallery.pyecharts.org/#/BMap/bmap_base
百度Geo和Map：可以制作到国家/区域/市/区县四级散点分布图，但是无法支持街道的详细地图
百度Bmap：是百度地图的在线画图，但是从别的地方搜来的POI（如高德）即便转换成百度坐标但也可能不存在该点而画不了

Baidu is shit!

"""
